requirements.txt
absl-py==2.2.2
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
geopandas==1.0.1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.9
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==4.25.7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.7
six==1.17.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
